moved queries to helpers, and moved the querys turnery to make two -complete- queries instead of modifying the content of a single query
